Move legacy controllers & transformers into Legacy namespace.
